Fix documentation links
also:
- rename seeds => menu
- update examples README
- stub new examples
