+++ crates/bevy_ui_navigation/src/events.rs
@@ -1,18 +1,21 @@
 //! Navigation events and requests.
 //!
-//! The navigation system works through bevy's `Events` system. Basically, it is
-//! a system with one input and two outputs:
-//! * Input [`Events<NavRequest>`], tells the navigation system what to do.
-//!   Your app should have a system that writes to a [`EventWriter<NavRequest>`]
-//!   based on inputs or internal game state. Usually, the default input systems specified
-//!   in [`crate::systems`] do that for you. But you can add your own requests
-//!   on top of the ones the default systems send. For example to unlock the UI with
-//!   [`NavRequest::Free`].
-//! * Output [`Focusable`](crate::Focusable) components. The navigation system
-//!   updates the focusables component according to the focus state of the
-//!   navigation system. See examples directory for how to read those
-//! * Output [`EventReader<NavEvent>`], contains specific information about what
-//!   the navigation system is doing.
+//! The navigation system works through bevy's `Events` system.
+//! It is a system with one input and two outputs:
+//! * Input `EventWriter<NavRequest>`, tells the navigation system what to do.
+//!   Your app should have a system that writes to a `EventWriter<NavRequest>`
+//!   based on inputs or internal game state.
+//!   Bevy provides default systems in `bevy_ui`.
+//!   But you can add your own requests on top of the ones the default systems send.
+//!   For example to unlock the UI with [`NavRequest::Free`].
+//! * Output [`Focusable`] components.
+//!   The navigation system updates the focusables component
+//!   according to the focus state of the navigation system.
+//!   See `examples/cursor_navigation` directory for usage clues.
+//! * Output `EventReader<NavEvent>`,
+//!   contains specific information about what the navigation system is doing.
+//!
+//! [`Focusable`]: crate::resolve::Focusable
 use bevy_ecs::{
     entity::Entity,
     event::EventReader,
@@ -24,16 +27,30 @@ use non_empty_vec::NonEmpty;
 /// Requests to send to the navigation system to update focus.
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub enum NavRequest {
-    /// Move in 2d in provided direction.
+    /// Move in in provided direction according to the plugin's [navigation strategy].
+    ///
+    /// Typically used by gamepads.
+    ///
+    /// [navigation strategy]: crate::resolve::MenuNavigationStrategy.
     Move(Direction),
-    /// Move within the encompassing [`NavMenu::BoundScope`](crate::NavMenu::BoundScope).
+    /// Move within the encompassing [`MenuSetting::scope`].
+    ///
+    /// [`MenuSetting::scope`]: crate::prelude::MenuSetting::scope
     ScopeMove(ScopeDirection),
-    /// Enter submenu if any [`NavMenu::reachable_from`](crate::NavMenu::reachable_from)
-    /// the currently focused entity.
+    /// Activate the currently focused [`Focusable`].
+    ///
+    /// If a menu is _[reachable from]_
+    ///
+    /// [`Focusable`]: crate::prelude::Focusable
+    /// [reachable from]: crate::menu::MenuBuilder::NamedParent
     Action,
-    /// Leave this submenu to enter the one it is [`reachable_from`](crate::NavMenu::reachable_from).
+    /// Leave this submenu to enter the one it is _[reachable from]_.
+    ///
+    /// [reachable from]: crate::menu::MenuBuilder::NamedParent
     Cancel,
-    /// Move the focus to any arbitrary [`Focusable`](crate::Focusable) entity.
+    /// Move the focus to any arbitrary [`Focusable`] entity.
+    ///
+    /// [`Focusable`]: crate::resolve::Focusable
     FocusOn(Entity),
     /// Unlocks the navigation system.
     ///
@@ -42,7 +59,9 @@ pub enum NavRequest {
     Free,
 }
 
-/// Direction for movement in [`NavMenu::BoundScope`](crate::NavMenu::BoundScope) menus.
+/// Direction for movement in [`MenuSetting::scope`] menus.
+///
+/// [`MenuSetting::scope`]: crate::menu::MenuSetting
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub enum ScopeDirection {
     Next,
@@ -68,7 +87,9 @@ pub enum NavEvent {
     /// Tells the app which element is the first one to be focused.
     ///
     /// This will be sent whenever the number of focused elements go from 0 to 1.
-    /// Meaning: whenever you spawn a new UI with [`crate::Focusable`] elements.
+    /// Meaning: whenever you spawn a new UI with [`Focusable`] elements.
+    ///
+    /// [`Focusable`]: crate::resolve::Focusable
     InitiallyFocused(Entity),
 
     /// Focus changed.
@@ -97,16 +118,20 @@ pub enum NavEvent {
         /// The [`NavRequest`] that didn't do anything.
         request: NavRequest,
     },
-    /// A [lock focusable](crate::Focusable::lock) has been triggered.
+    /// A [lock focusable] has been triggered.
     ///
     /// Once the navigation plugin enters a locked state, the only way to exit
     /// it is to send a [`NavRequest::Free`].
+    ///
+    /// [lock focusable]: crate::resolve::Focusable::lock
     Locked(Entity),
 
-    /// A [lock focusable](crate::Focusable::lock) has been released.
+    /// A [lock focusable] has been released.
     ///
     /// The navigation system was in a locked state triggered by `Entity`,
     /// is now unlocked, and receiving events again.
+    ///
+    /// [lock focusable]: crate::resolve::Focusable::lock
     Unlocked(Entity),
 }
 impl NavEvent {
@@ -120,7 +145,7 @@ impl NavEvent {
         }
     }
 
-    /// Whether this event is a [`NavEvent::NoChange`]
+    /// Whether this event is a [`NavEvent::NoChanges`]
     /// triggered by a [`NavRequest::Action`]
     /// if `entity` is the currently focused element.
     pub fn is_activated(&self, entity: Entity) -> bool {
@@ -167,6 +192,8 @@ impl<'w, 's, 'a> NavEventReader<'w, 's, 'a> {
     ///
     /// A [`Focusable`] is _activated_ when a [`NavRequest::Action`] is sent
     /// while it is focused, and it doesn't lead to a new menu.
+    ///
+    /// [`Focusable`]: crate::resolve::Focusable
     pub fn activated(&mut self) -> impl Iterator<Item = Entity> + '_ {
         self.with_request(NavRequest::Action)
     }
