+++ crates/bevy_ui_navigation/src/lib.rs
@@ -1,8 +1,8 @@
 mod commands;
 pub mod events;
+mod menu;
 mod named;
 mod resolve;
-mod seeds;
 
 use std::marker::PhantomData;
 
@@ -15,17 +15,12 @@ use bevy_ecs::{
 pub use non_empty_vec::NonEmpty;
 use resolve::TreeMenu;
 
-/// The [`Bundle`](bevy::prelude::Bundle)s
-/// returned by the [`NavMenu`] methods.
-pub mod bundles {
-    pub use crate::seeds::MenuBuilderBundle;
-}
 pub mod prelude {
     pub use crate::events::{NavEvent, NavEventReaderExt, NavRequest};
+    pub use crate::menu::{MenuBuilder, MenuSetting};
     pub use crate::resolve::{
         FocusAction, FocusState, Focusable, Focused, MenuNavigationStrategy, NavLock,
     };
-    pub use crate::seeds::{MenuBuilder, MenuSetting};
 }
 
 /// The label of the system in which the [`NavRequest`] events are handled, the
@@ -33,8 +28,8 @@ pub mod prelude {
 /// are sent.
 ///
 /// Systems updating visuals of UI elements should run _after_ the `NavRequestSystem`,
-/// while systems that emit [`NavRequest`] should run _before_ it. For example, the
-/// [`systems::default_mouse_input`] systems should run before the `NavRequestSystem`.
+/// while systems that emit [`NavRequest`] should run _before_ it.
+/// For example, an input system should run before the `NavRequestSystem`.
 ///
 /// Failing to do so won't cause logical errors, but will make the UI feel more slugish
 /// than necessary. This is especially critical of you are running on low framerate.
@@ -42,13 +37,21 @@ pub mod prelude {
 /// # Example
 ///
 /// ```rust, no_run
-/// use bevy::prelude::*;
-/// use bevy_ui_navigation::{NavRequestSystem, DefaultNavigationPlugins};
+/// use bevy_ui_navigation::prelude::*;
+/// # #[derive(SystemParam)] struct MoveCursor3d<'w, 's>(PhantomData<&'w &'s ()>);
+/// # impl<'w, 's> MenuNavigationStrategy for MoveCursor3d<'w, 's> {
+/// #   fn resolve_2d<'a>(
+/// #       &self,
+/// #       focused: Entity,
+/// #       direction: events::Direction,
+/// #       cycles: bool,
+/// #       siblings: &'a [Entity],
+/// #   ) -> Option<&'a Entity> { None }
+/// # }
 /// # fn button_system() {}
 /// fn main() {
 ///     App::new()
-///         .add_plugins(DefaultPlugins)
-///         .add_plugins(DefaultNavigationPlugins)
+///         .add_plugin(NavigationPlugin::<MoveCursor3d>::new())
 ///         // ...
 ///         // Add the button color update system after the focus update system
 ///         .add_system(button_system.after(NavRequestSystem))
@@ -56,32 +59,38 @@ pub mod prelude {
 ///         .run();
 /// }
 /// ```
+///
+/// [`NavRequest`]: prelude::NavRequest
+/// [`NavEvent`]: prelude::NavEvent
+/// [`Focusable`]: prelude::Focusable
 #[derive(Clone, Debug, Hash, PartialEq, Eq, SystemLabel)]
 pub struct NavRequestSystem;
 
 /// The navigation plugin.
 ///
-/// Add it to your app with `.add_plugin(NavigationPlugin)` and send
+/// Add it to your app with `.add_plugin(NavigationPlugin::new())` and send
 /// [`NavRequest`]s to move focus within declared [`Focusable`] entities.
 ///
-/// This means you'll also have to add manaully the systems from [`systems`]
-/// and [`systems::InputMapping`]. You should prefer [`DefaultNavigationPlugins`]
+/// You should prefer `bevy_ui` provided defaults
 /// if you don't want to bother with that.
 ///
 /// # Note on generic parameters
 ///
-/// The `MP` type parameter might seem complicated, but all you have to do
-/// is for your type to implement [`SystemParam`] and [`MoveParam`].
-/// See the [`resolve::UiProjectionQuery`] source code for implementation hints.
+/// The `STGY` type parameter might seem complicated, but all you have to do
+/// is for your type to implement [`SystemParam`] and [`MenuNavigationStrategy`].
+///
+/// [`MenuNavigationStrategy`]: resolve::MenuNavigationStrategy
+/// [`Focusable`]: prelude::Focusable
+/// [`NavRequest`]: prelude::NavRequest
 #[derive(Default)]
-pub struct GenericNavigationPlugin<MP>(PhantomData<fn() -> MP>);
+pub struct NavigationPlugin<STGY>(PhantomData<fn() -> STGY>);
 
-impl<STGY: resolve::MenuNavigationStrategy> GenericNavigationPlugin<STGY> {
+impl<STGY: resolve::MenuNavigationStrategy> NavigationPlugin<STGY> {
     pub fn new() -> Self {
         Self(PhantomData)
     }
 }
-impl<STGY: SystemParam + 'static> Plugin for GenericNavigationPlugin<STGY>
+impl<STGY: SystemParam + 'static> Plugin for NavigationPlugin<STGY>
 where
     for<'w, 's> SystemParamItem<'w, 's, STGY>: resolve::MenuNavigationStrategy,
 {
@@ -90,7 +99,7 @@ where
         app.register_type::<resolve::Focusable>()
             .register_type::<resolve::FocusState>()
             .register_type::<resolve::FocusAction>()
-            .register_type::<seeds::MenuSetting>()
+            .register_type::<menu::MenuSetting>()
             .register_type::<TreeMenu>();
 
         app.add_event::<events::NavRequest>()
